mastodon-instance/dummy-data-generator/requirements.txt
aiohttp==3.9.0
aiosignal==1.3.1
attrs==23.1.0
blurhash==1.1.4
certifi==2023.11.17
charset-normalizer==3.3.2
datasets==2.15.0
decorator==5.1.1
dill==0.3.7
Faker==20.1.0
filelock==3.13.1
frozenlist==1.4.0
fsspec==2023.10.0
huggingface-hub==0.19.4
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
Mastodon.py==1.8.1
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.0
numpy==1.26.2
packaging==23.2
pandas==2.1.3
Pillow==9.3.0
psycopg==3.1.13
psycopg-binary==3.1.13
pyarrow==14.0.1
pyarrow-hotfix==0.6
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
six==1.16.0
sympy==1.12
tokenizers==0.15.0
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
tqdm==4.66.1
transformers==4.35.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
xxhash==3.4.1
yarl==1.9.3
